The files of model : Admin and Payment was updated

+++ backendPY/models/payment.model.js
@@ -3,7 +3,7 @@ const {DataTypes} = require('sequelize');
 
 const Payment = sequelize.define("payment", {
     amount: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT(10, 2),
         allowNull: false
     }
     //THE ORDER ID CAN BE NULL FOR INITIALIZATION BEFORE
